Dwa endpointy + transakcja

+++ PrzykladoweKolokwium2025_1/PrzykladoweKolokwium2025_1/Models/DTOs/MowieDTO.cs
@@ -3,5 +3,5 @@ namespace PrzykladoweKolokwium2025_1.Models;
 public class MowieDTO
 {
     public string title { get; set; }
-    public double priceAtRental { get; set; }
+    public decimal priceAtRental { get; set; }
 }
